project1/requirements.txt
streamlit
Pillow
mutagen
numpy
